Create whiteimage.py
The provided Python code utilizes the OpenCV library to generate and save a pure white image with specified dimensions. The script begins by importing the necessary libraries, namely NumPy for numerical operations and OpenCV (cv2) for computer vision tasks. The dimensions of the desired image are then set to 840 pixels in width and 680 pixels in height.

Using NumPy, a white image is created as a three-dimensional array, where each pixel is assigned the maximum intensity value of 255 to represent the color white. The code proceeds to specify a file path ('save_path') where the resulting image will be saved. In this instance, the image is intended to be saved in the "C:\Users\Sumit Kumar\Desktop\opencv" directory with the filename "pure_white_image.png".

The script employs the 'cv2.imwrite' function to save the generated pure white image to the specified path. Additionally, the code includes an optional step to display the image using OpenCV's 'cv2.imshow' function, allowing for visual verification.

Lastly, the script includes commands to wait for a key press (cv2.waitKey(0)) to keep the image window open indefinitely and subsequently closes all OpenCV windows (cv2.destroyAllWindows()). In summary, the code serves to create, save, and optionally display a pure white image using OpenCV in a straightforward manner